requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
easyocr==1.7.2
filelock==3.19.1
fsspec==2025.7.0
ftfy==6.3.1
hf-xet==1.1.7
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
open_clip_torch==3.1.0
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
pyclipper==1.3.0.post6
python-bidi==0.6.6
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2025.7.34
requests==2.32.4
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.1
shapely==2.1.1
sympy==1.14.0
tifffile==2025.6.11
timm==1.0.19
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
